+++ scripts/test/Selenium/Agent/Admin/AdminGenericAgent.t
@@ -21,6 +21,9 @@ $Selenium->RunTest(
         # get helper object
         my $Helper = $Kernel::OM->Get('Kernel::System::UnitTest::Helper');
 
+        # get needed variable
+        my $RandomID = $Helper->GetRandomID();
+
         # set generic agent run limit
         $Helper->ConfigSettingChange(
             Valid => 1,
@@ -51,12 +54,38 @@ $Selenium->RunTest(
             UserLogin => $TestUserLogin,
         );
 
+        # get dynamic field object
+        my $DynamicFieldObject = $Kernel::OM->Get('Kernel::System::DynamicField');
+
+        # create test dynamic field of type date
+        my $DynamicFieldName = 'Test' . $RandomID;
+        my $DynamicFieldID   = $DynamicFieldObject->DynamicFieldAdd(
+            Name       => $DynamicFieldName,
+            Label      => $DynamicFieldName,
+            FieldOrder => 9991,
+            FieldType  => 'Date',
+            ObjectType => 'Ticket',
+            Config     => {
+                DefaultValue    => 0,
+                YearsInFuture   => 0,
+                YearsInPast     => 0,
+                YearsPeriod     => 0,
+                DateRestriction => 'DisablePastDates',    # turn on validation of no past dates
+            },
+            ValidID => 1,
+            UserID  => $UserID,
+        );
+
+        $Self->True(
+            $DynamicFieldID,
+            "Dynamic field $DynamicFieldName - ID $DynamicFieldID - created",
+        );
+
         # get ticket object
         my $TicketObject = $Kernel::OM->Get('Kernel::System::Ticket');
 
         # create test tickets
-        my $TestTicketRandomID = $Helper->GetRandomID();
-        my $TestTicketTitle    = "Test Ticket $TestTicketRandomID Generic Agent";
+        my $TestTicketTitle = "Test Ticket $RandomID Generic Agent";
         my @TicketNumbers;
         for ( 1 .. 20 ) {
 
@@ -108,20 +137,25 @@ $Selenium->RunTest(
         $Selenium->execute_script('$(".WidgetSimple.Collapsed .WidgetAction.Toggle a").click();');
 
         # create test job
-        my $GenericTicketSearch = "*Ticket $TestTicketRandomID Generic*";
-        my $RandomID            = "GenericAgent" . $Helper->GetRandomID();
-        $Selenium->find_element( "#Profile", 'css' )->send_keys($RandomID);
+        my $GenericTicketSearch = "*Ticket $RandomID Generic*";
+        my $GenericAgentJob     = "GenericAgent" . $RandomID;
+        $Selenium->find_element( "#Profile", 'css' )->send_keys($GenericAgentJob);
         $Selenium->find_element( "#Title",   'css' )->send_keys($GenericTicketSearch);
+
+        # set test dynamic field to past date (bug#12210)
+        $Selenium->find_element( "#DynamicField_${DynamicFieldName}Year", 'css' )->send_keys('2015');
+
+        # save job
         $Selenium->find_element( "#Profile", 'css' )->VerifiedSubmit();
 
         # check if test job show on AdminGenericAgent
         $Self->True(
-            index( $Selenium->get_page_source(), $RandomID ) > -1,
-            "$RandomID job found on page",
+            index( $Selenium->get_page_source(), $GenericAgentJob ) > -1,
+            "$GenericAgentJob job found on page",
         );
 
         # edit test job to delete test ticket
-        $Selenium->find_element( $RandomID, 'link_text' )->VerifiedClick();
+        $Selenium->find_element( $GenericAgentJob, 'link_text' )->VerifiedClick();
 
         # toggle Execute Ticket Commands widget
         $Selenium->execute_script('$(".WidgetSimple.Collapsed .WidgetAction.Toggle a").click();');
@@ -129,7 +163,7 @@ $Selenium->RunTest(
         $Selenium->find_element( "#Profile", 'css' )->VerifiedSubmit();
 
         # run test job
-        $Selenium->find_element("//a[contains(\@href, \'Subaction=Run;Profile=$RandomID' )]")->VerifiedClick();
+        $Selenium->find_element("//a[contains(\@href, \'Subaction=Run;Profile=$GenericAgentJob' )]")->VerifiedClick();
 
         # verify there are no tickets found with enabled ExtendedSearchCondition
         $Self->True(
@@ -148,7 +182,7 @@ $Selenium->RunTest(
         $Selenium->VerifiedGet("${ScriptAlias}index.pl?Action=AdminGenericAgent");
 
         # run test job
-        $Selenium->find_element("//a[contains(\@href, \'Subaction=Run;Profile=$RandomID' )]")->VerifiedClick();
+        $Selenium->find_element("//a[contains(\@href, \'Subaction=Run;Profile=$GenericAgentJob' )]")->VerifiedClick();
 
         # check if test job show expected result
         for my $TicketNumber (@TicketNumbers) {
@@ -173,7 +207,7 @@ $Selenium->RunTest(
         $Selenium->find_element("//a[contains(\@href, \'Subaction=RunNow' )]")->VerifiedClick();
 
         # run test job again
-        $Selenium->find_element("//a[contains(\@href, \'Subaction=Run;Profile=$RandomID' )]")->VerifiedClick();
+        $Selenium->find_element("//a[contains(\@href, \'Subaction=Run;Profile=$GenericAgentJob' )]")->VerifiedClick();
 
         # check if there is no warning message:
         # "Affected more tickets than how many will be executed on run job"
@@ -188,7 +222,7 @@ $Selenium->RunTest(
         $Selenium->find_element("//a[contains(\@href, \'Subaction=RunNow' )]")->VerifiedClick();
 
         # set test job to invalid
-        $Selenium->find_element( $RandomID, 'link_text' )->VerifiedClick();
+        $Selenium->find_element( $GenericAgentJob, 'link_text' )->VerifiedClick();
 
         $Selenium->execute_script("\$('#Valid').val('0').trigger('redraw.InputField').trigger('change');");
         $Selenium->find_element( "#Profile", 'css' )->VerifiedSubmit();
@@ -196,15 +230,25 @@ $Selenium->RunTest(
         # check class of invalid generic job in the overview table
         $Self->True(
             $Selenium->execute_script(
-                "return \$('tr.Invalid td:contains($RandomID)').length"
+                "return \$('tr.Invalid td:contains($GenericAgentJob)').length"
             ),
             "There is a class 'Invalid' for test generic job",
         );
 
         # delete test job
-        $Selenium->find_element("//a[contains(\@href, \'Subaction=Delete;Profile=$RandomID\' )]")->VerifiedClick();
+        $Selenium->find_element("//a[contains(\@href, \'Subaction=Delete;Profile=$GenericAgentJob\' )]")
+            ->VerifiedClick();
 
-    }
+        # delete created test dynamic field
+        my $Success = $DynamicFieldObject->DynamicFieldDelete(
+            ID     => $DynamicFieldID,
+            UserID => $UserID,
+        );
+        $Self->True(
+            $Success,
+            "Dynamic field - ID $DynamicFieldID - deleted",
+        );
+    },
 
 );